style(dataset-compare-ui): move compare view charts to tabs
